Share EffectFlags
