+++ CSSE/CSMagicEffect.cpp
@@ -73,7 +73,7 @@ namespace se::cs {
 	}
 
 	int MagicEffect::getNameGMST() const {
-		if (id < EffectID::FirstEffect || id > EffectID::LastEffect) {
+		if (id < TES3::EffectID::FirstEffect || id > TES3::EffectID::LastEffect) {
 			return -1;
 		}
 		const auto gConvertEffectToGMST = reinterpret_cast<int*>(0x6A7E74);
@@ -101,139 +101,139 @@ namespace se::cs {
 	}
 
 	bool MagicEffect::getFlagTargetSkill() const {
-		return getEffectFlag(EffectFlag::TargetSkillBit);
+		return getEffectFlag(TES3::EffectFlag::TargetSkillBit);
 	}
 
 	void MagicEffect::setFlagTargetSkill(bool value) const {
-		setEffectFlag(EffectFlag::TargetSkillBit, value);
+		setEffectFlag(TES3::EffectFlag::TargetSkillBit, value);
 	}
 
 	bool MagicEffect::getFlagTargetAttribute() const {
-		return getEffectFlag(EffectFlag::TargetAttributeBit);
+		return getEffectFlag(TES3::EffectFlag::TargetAttributeBit);
 	}
 
 	void MagicEffect::setFlagTargetAttribute(bool value) const {
-		setEffectFlag(EffectFlag::TargetAttributeBit, value);
+		setEffectFlag(TES3::EffectFlag::TargetAttributeBit, value);
 	}
 
 	bool MagicEffect::getFlagNoDuration() const {
-		return getEffectFlag(EffectFlag::NoDurationBit);
+		return getEffectFlag(TES3::EffectFlag::NoDurationBit);
 	}
 
 	void MagicEffect::setFlagNoDuration(bool value) const {
-		setEffectFlag(EffectFlag::NoDurationBit, value);
+		setEffectFlag(TES3::EffectFlag::NoDurationBit, value);
 	}
 
 	bool MagicEffect::getFlagNoMagnitude() const {
-		return getEffectFlag(EffectFlag::NoMagnitudeBit);
+		return getEffectFlag(TES3::EffectFlag::NoMagnitudeBit);
 	}
 
 	void MagicEffect::setFlagNoMagnitude(bool value) const {
-		setEffectFlag(EffectFlag::NoMagnitudeBit, value);
+		setEffectFlag(TES3::EffectFlag::NoMagnitudeBit, value);
 	}
 
 	bool MagicEffect::getFlagHarmful() const {
-		return getEffectFlag(EffectFlag::HarmfulBit);
+		return getEffectFlag(TES3::EffectFlag::HarmfulBit);
 	}
 
 	void MagicEffect::setFlagHarmful(bool value) const {
-		setEffectFlag(EffectFlag::HarmfulBit, value);
+		setEffectFlag(TES3::EffectFlag::HarmfulBit, value);
 	}
 
 	bool MagicEffect::getFlagContinuousVFX() const {
-		return getEffectFlag(EffectFlag::ContinuousVFXBit);
+		return getEffectFlag(TES3::EffectFlag::ContinuousVFXBit);
 	}
 
 	void MagicEffect::setFlagContinuousVFX(bool value) const {
-		setEffectFlag(EffectFlag::ContinuousVFXBit, value);
+		setEffectFlag(TES3::EffectFlag::ContinuousVFXBit, value);
 	}
 
 	bool MagicEffect::getFlagCanCastSelf() const {
-		return getEffectFlag(EffectFlag::CanCastSelfBit);
+		return getEffectFlag(TES3::EffectFlag::CanCastSelfBit);
 	}
 
 	void MagicEffect::setFlagCanCastSelf(bool value) const {
-		setEffectFlag(EffectFlag::CanCastSelfBit, value);
+		setEffectFlag(TES3::EffectFlag::CanCastSelfBit, value);
 	}
 
 	bool MagicEffect::getFlagCanCastTouch() const {
-		return getEffectFlag(EffectFlag::CanCastTouchBit);
+		return getEffectFlag(TES3::EffectFlag::CanCastTouchBit);
 	}
 
 	void MagicEffect::setFlagCanCastTouch(bool value) const {
-		setEffectFlag(EffectFlag::CanCastTouchBit, value);
+		setEffectFlag(TES3::EffectFlag::CanCastTouchBit, value);
 	}
 
 	bool MagicEffect::getFlagCanCastTarget() const {
-		return getEffectFlag(EffectFlag::CanCastTargetBit);
+		return getEffectFlag(TES3::EffectFlag::CanCastTargetBit);
 	}
 
 	void MagicEffect::setFlagCanCastTarget(bool value) const {
-		setEffectFlag(EffectFlag::CanCastTargetBit, value);
+		setEffectFlag(TES3::EffectFlag::CanCastTargetBit, value);
 	}
 
 	bool MagicEffect::getFlagNegativeLighting() const {
-		return getEffectFlag(EffectFlag::NegativeLightingBit);
+		return getEffectFlag(TES3::EffectFlag::NegativeLightingBit);
 	}
 
 	void MagicEffect::setFlagNegativeLighting(bool value) const {
-		setEffectFlag(EffectFlag::NegativeLightingBit, value);
+		setEffectFlag(TES3::EffectFlag::NegativeLightingBit, value);
 	}
 
 	bool MagicEffect::getFlagAppliedOnce() const {
-		return getEffectFlag(EffectFlag::AppliedOnceBit);
+		return getEffectFlag(TES3::EffectFlag::AppliedOnceBit);
 	}
 
 	void MagicEffect::setFlagAppliedOnce(bool value) const {
-		setEffectFlag(EffectFlag::AppliedOnceBit, value);
+		setEffectFlag(TES3::EffectFlag::AppliedOnceBit, value);
 	}
 
 	bool MagicEffect::getFlagNonRecastable() const {
-		return getEffectFlag(EffectFlag::NonRecastableBit);
+		return getEffectFlag(TES3::EffectFlag::NonRecastableBit);
 	}
 
 	void MagicEffect::setFlagNonRecastable(bool value) const {
-		setEffectFlag(EffectFlag::NonRecastableBit, value);
+		setEffectFlag(TES3::EffectFlag::NonRecastableBit, value);
 	}
 
 	bool MagicEffect::getFlagIllegalDaedra() const {
-		return getEffectFlag(EffectFlag::IllegalDaedraBit);
+		return getEffectFlag(TES3::EffectFlag::IllegalDaedraBit);
 	}
 
 	void MagicEffect::setFlagIllegalDaedra(bool value) const {
-		setEffectFlag(EffectFlag::IllegalDaedraBit, value);
+		setEffectFlag(TES3::EffectFlag::IllegalDaedraBit, value);
 	}
 
 	bool MagicEffect::getFlagUnreflectable() const {
-		return getEffectFlag(EffectFlag::UnreflectableBit);
+		return getEffectFlag(TES3::EffectFlag::UnreflectableBit);
 	}
 
 	void MagicEffect::setFlagUnreflectable(bool value) const {
-		setEffectFlag(EffectFlag::UnreflectableBit, value);
+		setEffectFlag(TES3::EffectFlag::UnreflectableBit, value);
 	}
 
 	bool MagicEffect::getFlagCasterLinked() const {
-		return getEffectFlag(EffectFlag::CasterLinkedBit);
+		return getEffectFlag(TES3::EffectFlag::CasterLinkedBit);
 	}
 
 	void MagicEffect::setFlagCasterLinked(bool value) const {
-		setEffectFlag(EffectFlag::CasterLinkedBit, value);
+		setEffectFlag(TES3::EffectFlag::CasterLinkedBit, value);
 	}
 
 	bool MagicEffect::getAllowSpellmaking() const {
-		return BIT_TEST(flags, EffectFlag::AllowSpellmakingBit);
+		return BIT_TEST(flags, TES3::EffectFlag::AllowSpellmakingBit);
 	}
 
 	void MagicEffect::setAllowSpellmaking(bool value) {
-		BIT_SET(flags, EffectFlag::AllowSpellmakingBit, value);
+		BIT_SET(flags, TES3::EffectFlag::AllowSpellmakingBit, value);
 	}
 
 	bool MagicEffect::getAllowEnchanting() const {
-		return BIT_TEST(flags, EffectFlag::AllowEnchantingBit);
+		return BIT_TEST(flags, TES3::EffectFlag::AllowEnchantingBit);
 	}
 
 	void MagicEffect::setAllowEnchanting(bool value) {
-		BIT_SET(flags, EffectFlag::AllowEnchantingBit, value);
+		BIT_SET(flags, TES3::EffectFlag::AllowEnchantingBit, value);
 	}
 
 	bool MagicEffect::search(std::string_view needle, const BaseObject::SearchSettings& settings, std::regex* regex, int attribute, int skill) const {
