+++ SharedSE/SharedSE.vcxitems
@@ -118,6 +118,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)Deque.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)StdString.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3AttachmentFlags.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3EffectFlags.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3ObjectFlags.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3ObjectType.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3RaceFlags.h" />
